+++ lib/constants.js
@@ -55,10 +55,83 @@ const transport = nodemailer.createTransport({
     }
   });
 
+const ascReturnResult = {
+    "rows": [
+        {
+            "name": "test001",
+            "email": "test001@vanderbilt.edu",
+            "visions": 1,
+            "type": "vuceptor",
+            "status": "unregistered"
+        },
+        {
+            "name": "test002",
+            "email": "test002@vanderbilt.edu",
+            "visions": 2,
+            "type": "vuceptor",
+            "status": "unregistered"
+        },
+        {
+            "name": "test003",
+            "email": "test003@vanderbilt.edu",
+            "visions": 3,
+            "type": "vuceptor",
+            "status": "unregistered"
+        },
+        {
+            "name": "test005",
+            "email": "test005@vanderbilt.edu",
+            "visions": 5,
+            "type": "vuceptor",
+            "status": "unregistered"
+        }
+    ],
+    "pages": 1
+  };
+
+const descReturnResult = {
+    "rows": [
+        {
+            "name": "test005",
+            "email": "test005@vanderbilt.edu",
+            "visions": 5,
+            "type": "vuceptor",
+            "status": "unregistered"
+        },
+        {
+            "name": "test003",
+            "email": "test003@vanderbilt.edu",
+            "visions": 3,
+            "type": "vuceptor",
+            "status": "unregistered"
+        },
+        {
+            "name": "test002",
+            "email": "test002@vanderbilt.edu",
+            "visions": 2,
+            "type": "vuceptor",
+            "status": "unregistered"
+        },
+        {
+            "name": "test001",
+            "email": "test001@vanderbilt.edu",
+            "visions": 1,
+            "type": "vuceptor",
+            "status": "unregistered"
+        }
+    ],
+    "pages": 1
+  };
+
+const emptyReturnResult = {"rows": [], "pages": 0};
+
 module.exports = {
     transport,
     STATUS_CODE,
     SORT_ORDER,
     TYPE,
-    REGISTRATION_STATUS
+    REGISTRATION_STATUS,
+    ascReturnResult,
+    descReturnResult,
+    emptyReturnResult
 }
