+++ routes/user.js
@@ -8,7 +8,6 @@ routerUser.post('/userLoadfromcsv', userController.userLoadfromcsv);
 routerUser.post('/createUser', userController.createUser);
 routerUser.post('/updateUser', userController.updateUser);
 routerUser.post('/deleteUser', userController.deleteUser);
-
 routerUser.get('/readUser', userController.readUser);
 routerUser.get('/visionsNums', userController.visionsNums);
